Preload the assets with an empty buffer scene
Published builds automatically preload all assets in a scene when the
scene loads. The exception to this rule is scene 0. This is because the
first scene is usually a splashscreen, which you want to display as
quickly as possible.

That's why we add a preloading scene as scene 0 to act as a buffer to
preload the assets of the first next scene.